refactor : 요구사항에 맞춰 메서드 분리
